requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
authlib==1.3.1
google-auth==2.29.0
aiosmtplib==2.0.2
requests==2.31.0
edge-tts==7.0.2
pydub==0.25.1
httpx==0.27.0
pytest==8.3.2
